pve, common: start 'bindings' module and documentation
In order to have better documentation of all the packages provided by
the rust code, we introduce a new convention of how to organize the
code.

It is described in the README.md.

Essentially, there's a `bindings` submodule with
`#![deny(missing_docs)]` in which every perl package should be
exported with its perl package name converted to snake_case (with
the double-colons replaced by single underscores), and all `#[export]`
functions should be `pub`.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>