pve-rs: remove PVE::RS::APT:Repositories module
We can remove the PVE::RS::APT::Repositories module because we needed
its functionality in Proxmox VE and other products, so we moved it to
Proxmox::RS::APT::Repositories some time ago. We now only use the new
module in our code.
When users update to Proxmox VE 9, we can be sure that there is no
package using PVE::RS::APT::Repositories, so we can safely remove the
code.

Signed-off-by: Hannes Duerr <h.duerr@proxmox.com>