pmg-rs: remove PMG::RS::APT:Repositories module
We can remove the PMG::RS::APT::Repositories module because we needed
its functionality in Proxmox Mail Gateway (PMG) and other products, so
we moved it to Proxmox::RS::APT::Repositories some time ago. We now only
use the new module in our code.
When users update to PMG 9, we can be sure that there is no
package using PMG::RS::APT::Repositories, so we can safely remove the
code.

Signed-off-by: Hannes Duerr <h.duerr@proxmox.com>