pve-rs: Add PVE::RS::SDN::Fabrics module
This module exposes the functionality provided proxmox-ve-config for
the SDN fabrics to perl. We add initial support for reading and
writing the section config stored in /etc/pve/sdn/fabrics.cfg as well
as the running configuration, stored in /etc/pve/sdn/.running-config.
It also provides a helper method for calculating the digest of the
configuration.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-29-g.goller@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>