pve-rs: sdn: fabrics: add helper to generate ifupdown2 configuration
SDN fabrics can be used to configure IP addresses on interfaces
directly, so we need to generate the respective ifupdown2
configuration from the fabrics configuration. We also set some
additional properties that are required for interfaces that are part
of a fabric (IP forwarding). We use dummy interfaces, instead of
loopback interfaces, for configuring the router IP of the node, so for
each fabric we generate a dummy interface that carries the IP.

Currently this is a simple implementation that builds a String from
the SDN fabrics configuration, but in the future we intend to create a
full-fledged crate for reading / writing ifupdown2 configuration
files.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-32-g.goller@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>