pve-rs: sdn: fabrics: add helper for network API endpoint
In PVE we use the GET /nodes/{node}/network API endpoint to return all
currently configured network interfaces on a specific node. In order
to be able to use SDN fabrics in Ceph and the migration settings, we
add a helper method that returns all fabrics formatted in the same
format as the pre-existing PVE API call. This enables us to return the
SDN fabrics in the endpoint so users can select the fabrics from the
UI, integrating the fabrics with the existing UI components.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-33-g.goller@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>