pve-rs: fabrics: add function to get l2vpn and l3vpn routes for evpn
Add functions to get the l3vpn and l2vpn routes from frr so that we can
display them in the ui in the evpn zone content view. L3vpn route
retrieval is quite easy, we just get the routes that are in a specific
vrf (the vrf of the zone) (we could do this with iproute2, but we use
vtysh so that we can get all nexthops). For l2vpn we could also use
`bridge fdb`, but then we don't get the VNI and the ip address
associated to the l2vpn route distributed by EVPN. In order to get all
the information we show all the type2 routes that EVPN receives and get
the mac and ip address of them. We also filter by installed and bestpath
so we only display the installed and best routes.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251107143201.689035-17-s.hanreich@proxmox.com
Tested-by: Hannes Dürr <h.duerr@proxmox.com>