pve-rs: fabrics: add function to get the interfaces used for a fabric
Add a function which queries FRR to return all the interfaces which are
used on the local node by a specific fabric. Again, we can associate a
fabric with a specific openfabric/ospf interface by reading and parsing
the fabric config and getting all the interfaces configured on the node
and matching them to the FRR output.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251107143201.689035-19-s.hanreich@proxmox.com
Tested-by: Hannes Dürr <h.duerr@proxmox.com>