pve-rs: fabrics: add function to get the neighbors for a fabric
Add a function that takes a fabricId and returns all routing protocol
neighbors (OpenFabric or OSPF) identified by FRR. For OpenFabric, the
fabric name is used as the area identifier, so we can filter easily. For
OSPF, the fabric configuration is read/parsed to extract interfaces,
which are then matched against the FRR neighbor output. The fabric
configuration is always read first to verify that a fabric exists before
querying FRR, as configuration parsing is (probably) faster than vtysh
execution.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251107143201.689035-20-s.hanreich@proxmox.com
Tested-by: Hannes Dürr <h.duerr@proxmox.com>