pve-rs: fabrics: add unit-tests for fabrics
Add unit-tests for fabrics that provide a frr output and check the
parsed struct that is then returned from the pve-network api. In order
to test the returned structs PartialEq and Eq have been derived on a few
structs and the hostname retrieval has been pulled out to mock the
functions.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251107143201.689035-21-s.hanreich@proxmox.com
Tested-by: Hannes Dürr <h.duerr@proxmox.com>